chore(frontend): apply review fixes (dead props, helper consolidation)
Address Opus reviewer pass on the point 3 split:

- Consolidate four helpers (formatDateTime, basenameFromPath,
  fileTypeFromPath, formatSizeMb) into a single src/lib/format.ts.
  AGENTS.md §3 ("Three similar lines is fine. Abstract on the fourth
  instance.") — they had been duplicated 3+ times across App.tsx,
  features/file-detail/utils.ts, features/duplicates/DuplicatesView,
  and features/favorites-sauce/SauceFavoritesSettings.
- features/file-detail/utils.ts is now a thin re-export so existing
  imports keep working.
- Drop dead props (AGENTS.md §17):
  - SauceFavoritesSettings: `sauceSettings` and `canonicalizeSauceKey`
    were declared but never used inside the component.
  - GalleryView: `galleryTotal` and `manualOrderState` were declared
    but never destructured.
  Removed from both the props interfaces and the call sites in
  App.tsx.
- Drop the `renderNeighborPreview` render-prop pass-through. App.tsx
  was wiring it as `(file, direction) => <FileDetailPreview ... />`,
  which is a needless indirection — FileDetailPanel now imports
  FileDetailPreview directly and renders it inline. AGENTS.md §3
  ("Do not introduce a new layer ... unless the problem cannot be
  expressed with what is already here.").
- Drop the `formatDateTime` prop from FileDetailPanel for the same
  reason — the helper is now a shared import.

Build + vitest + Playwright remain green. JS bundle 348KB → 348KB
(no change; tree-shaking already deduped helpers).

Refs #103
